requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==3.3
mysqlclient==2.1.0
pycparser==2.21
pyOpenSSL==21.0.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
Werkzeug==2.0.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML~=6.0